[new-generics] Handle refinements by sealing/unsealing GenericTs
Summary:
This diff allows refinements to work on generics by refining the upper bound of the generic, while maintaining it as a generic. For example,

```
function f<X: mixed>(x: X) {
  if (typeof x === 'number') {
    (x: X);
    (x: number);
    (x: string); // nope
  }
}
```
After the `typeof` refinement, `x` becomes a `GenericT` whose upper bound is `number` rather than `mixed`. It can still flow into `X`, but it can also be used as a number.

This diff introduces a general approach to performing operations (like refinement) on the upper bounds of `GenericT`s without having to special-case or duplicate code. It does so with a new `SealGenericT` use-type, which contains a generic ID and an output continuation bound (usually used as a `t_out`, but allows for operations on upper bounds later in the stack), Whenever a type flows into a `SealGenericT`, a `GenericT` with that type as its upper bound is created and then passed into the continuation. This approach will be used in a lot of places in this project.

This diff also adds a main function for handling generics as lower bounds. Similar to `any_propagation`, `handle_generic` is called at the beginning of `__flow`, and if it returns true, it bypasses the main `__flow` match statement.

Reviewed By: panagosg7

Differential Revision: D22517822

fbshipit-source-id: 0f9afc94387d92f66b0769b687bce0cf01ba0272