[enums] Parsing of unknown members (`...`)
Summary:
This adds parsing support for Flow Enums with unknown members.

The error for using this when the option is not enabled is part of the next diff.

Whether an enum has unknown members is logically part of the enum body, so I've put it as part of this AST node, though this does cause a bit of repetition.

Syntax:

```
enum E {
  A,
  B,
  ...
}
```
The `...` must be last, and cannot have a trialing comma.

Reviewed By: Hans-Halverson

Differential Revision: D24152192

fbshipit-source-id: a7194ae00d3971146e965a35a132fe050f3a06a6