Use WorkerCancel.with_no_cancellations throughout parsing_heaps.ml
Summary:
I don't think that this entirely ensures that all the operations we expect to be atomic actually are, but it's a start. For example, we add the file hash separately. What happens if we add the hash but do not end up adding the AST due to a cancellation? Does that lead to an acceptable state? I bet there are other similar ways to get into an inconsistent state due to cancellation.

I believe these operations take place in the main process, which does not get cancelled by the mechanism that cancels the worker processes. However, it seems prudent to protect them nonetheless.

Reviewed By: samwgoldman

Differential Revision: D24145621

fbshipit-source-id: d2344c262851dc59b5d6c9145951df7fe8502a35