prevent infinte recursion in object rest destructuring
Summary: When an `ObjRestT`'s `tout` has a copy of itself (or the `DestructuringT` that creates the `ObjRestT`) as its own upper bound, we can get into infinite loops. Normally we prevent the same type from flowing to the same `use_t` more than once with the Flow cache, but because `ObjRestT` flows a fresh object to its tout every time it is triggered, this mechanism fails us. Instead, we make use of the `ConstFoldExpansion` mechanism that is used to solve this same issue elsewhere in `flow_js`.

Reviewed By: panagosg7

Differential Revision: D24225015

fbshipit-source-id: 9f26eb484a979ca3e8315f0a2fefad5c33cf4f2b