Use public API from sharedmem tests
Summary:
Before this change, the hashtbl test created its own extern bindings to the
hh_shared C API. This makes it fragile and annoying to actually change the
underlying C API.

Instead, we should be testing the public API of the unit under test. I have made
the necessary changes to retain the spirit of the existing tests.

Reviewed By: jbrown215

Differential Revision: D24229512

fbshipit-source-id: 616dd5d4ef9acda3192adce02483e23dff7b0c23