[new-generics] Use GenericT to represent objects with spreaded generics
Summary:
This diff adds support for spreading `GenericT`s into objects.

An object with a generic spread into it behaves kind of like the a generic itself. Consider:
```
function f<X: {a: number}>(x: X): {...X} {
   var ret: {...X} = {...x};
   (ret.a: number);
   return ret;
}
```
When we spread `x` into an object, the resulting object has `a: number` as a property, due to the presence of `{a: number}` as an upper bound on `X`. However, it's clearly not *just* that object: it would be super unsound to allow `return {a: 42}` in this function, because the caller could have instantiated `X` with `{a: number, b:string}` and expect the result of the call to have a `b` property as well. On the other hand, since we have a level of minimum information about `{...X}` from `X`'s upper bound, we can safely read `.a` from it.

This behavior is, naturally, similar to the behavior of `X` itself: we can use it as a lower bound in constraints based on its definition-site upper bound, but we can't flow anything into it except itself--we can't manufacture new values of type `{...X}` without spreading a value of type `X`.

Because of this similarity, this diff uses `GenericT`s to represent objects that have had generics spread into them. We distinguish a "bare" generic `GenericT` like `X` from a spreaded `GenericT` like `{...X}` by including a distinction in the generic ID, which is now a more complicated data structure representing both bounds (bare generics) and spreads. With this new structure, the representation of the type `{...X}` above becomes
```
GenericT { bound = ObjT { a: number }, id = Spread [X] }
```
Compare this with the representation of `X`:
```
GenericT { bound = ObjT { a: number }, id = Bound X }
```
In both cases, the type represents something that will behave like `ObjT { a: number }` as a lower bound, but which is sealed by `X` (but sealed in different ways).

Unlike regular bounds, spreads can contain multiple generics: for example,
```
function f<X: {}, Y: {}>(x: X, y: Y): {...X, ...Y} {
   return {...x, ...y}
}
```
Here, the type of the return in this function is `GenericT { bound = ObjT {}, id = Spread [Y,X] }`. Multiple spreads are encoded as part of a list; this allows this diff to close a number of soundness holes that currently exist in Flow: https://fburl.com/htu3e7h6 is now (correctly) an error. The `satisfies` function in `generic.ml` handles checking whether a lower bound generic ID satisfies an upper bound; see the comments on that function for details.

Just like for objects themselves, the representation for generic spreads is different in the `object_kit` representation. Here, we add a new `generics` field to the `object_slice` representation, which represents the generic spreads included within an object. It's in this representation that the spreads are extended or subtracted (with ObjRest). The conversion to and from the object slice representation handles decomposing or reconstructing `GenericT`s appropriately.

Reviewed By: samwgoldman

Differential Revision: D23039247

fbshipit-source-id: 524007ba650054c411e760aa62b85502a4baf0ad