[new-generics] Split the internal (generic-ish) and external view of ThisClassT
Summary:
Currently the `this` type inside of a class definition is treated like a generic whose upper bound is the type of the class's instance--when class definitions are analyzed, `this` is added to the set of generics that are fed into generate-tests. Then the class itself is represented using `ThisClassT`, which is unified with the tvar representing the type of the class as a whole, which is "fixed" into `ClassT` when it's used internally or externally.

Given that, consider how this program is checked using generate-tests:

```
class C {
  +f: () => this;
  g(): this {
    return this.f();
  }
}
```
When `this=empty`, `this.f()` never actually fires, because `EmptyT ~> MethodT` is short-circuited. Then, when `this=C`, `this.f()` produces `C`, which flows into `this` successfully.

Now consider this with new generics (before this diff). When `this.f()` fires, the upper bound of the `GenericT`, `C`, flows into `MethodT`. This looks up `f` on the class `C`, whose type is `() => C`. We then flow `C` to the return type of `g`, which is a `GenericT`--and this causes an error, since you can't flow non-generic types into a `GenericT`.

This diff fixes this issue by slightly changing how class's self-references, using `ThisClassT`, are represented internally versus externally. Now, the `ThisClassT` type has an additional boolean field representing whether it's an internal use of `C` or an external use. When an internal `ThisClassT` is fixed, it is wrapped in a `GenericT`, because it can still represent the range of types with `C` as upper bound. (The external view of a `ThisClassT` remains the same.)

The upshot of this for the above example is that `this.f()` now produces a `GenericT` rather than an `InstanceT`, and so the above example typechecks.

Reviewed By: panagosg7

Differential Revision: D23200210

fbshipit-source-id: f30d45c445ebd9cc076c0d1d9fb1f9fdb31b8dfd