[new-generics] Introduce bespoke generic ID
Summary:
This diff replaces the ALoc.id that was used as the id for `GenericT`s with a bespoke structure that carries that information and more. Currently, the biggest change that this makes is that `GenericT`s no longer have to be directly nested: if we have a function definition like

`function f<X: number, Y:X>(y: Y) ... `

then `y` is now *not* a `GenericT` whose bound is a `GenericT`, but rather a `GenericT` whose bound is `number`, but which encodes in its id that it is bounded by `X` as well as defined by `Y`.

This allows us to "peek" directly inside a `GenericT`'s bounds without having to loop through any number of nested generics to find a "concrete" bound. In the future, this will also be core to the representation of spreaded generics.**

When a `GenericT` flows into another `GenericT`, we now use the `satisfies` function to determine if the id on the lower bound matches the id on the upper bound. This can now return a non-binary result: it's possible for the comparison to result in an unsatisfied id on the upper bound, or a "left over" id on the lower bound (due to unions). This is discussed in detail in the comments of `generic.ml`.

** preview for a later diff coming soon: the type of `{...X}` is a `GenericT` whose id is `Spread(X)` and whose bound is an object containing `X`'s upper bound's properties.

Reviewed By: samwgoldman

Differential Revision: D22656427

fbshipit-source-id: 60516654012ca81cafa61cf656b7293c16de9410