add error disallowing this annotations unless specifically enabled
Summary: Even though we have added parsing support for `this` annotations, we don't actually want people using them until babel and parts of the toolchain have been updated to support them. Until these changes are properly rolled out to all parts of the toolchain, we will emit a type error on any uses of this new feature unless it is opted into explicitly.

Reviewed By: gkz

Differential Revision: D24194631

fbshipit-source-id: 2c0d73332057e1856842a31b7b2fc823129dfbad