[new-generics] Support for computed generic keys
Summary:
This diff allows computed keys that are generics to work properly, by having the functions in question accept generic keys by looking at their upper bounds. We can't just get rid of the generic before this occurs, because it's important to be able to compare the genericness of the incoming key with the indexer key in cases like
```
function f<X: number>(x: X, o: {[X]: string}) { o[x] = 'a' }
```
This means that we need to include `GenericT` in the pattern matches for `writelike_obj_prop` and `read_obj_prop`.

Reviewed By: panagosg7

Differential Revision: D23253478

fbshipit-source-id: fa13d7660919f2593f58a1b06e5fe779cf03a3ef