[new-generics] Avoid wildcards with generic LHS in flow match, fix binary operations, and support matching props.
Summary:
Some use_types have wildcards on the left in the `__flow` match statement, which leads to issues if the LHS type is a generic (they don't know what to do with a GenericT). This diff flows the upper bound of the generic to those use_types. This is what would normally happen were it not for the wildcard in the match statement, above the part of the match statement that handles GenericTs.

This diff also fixes the binary operation helper functions to know that a GenericT on the RHS needs to be resolved/"flipped."

Finally, this diff allows `MatchingPropT ~> GenericT (bound)` flows to turn into `MatchingPropT ~> bound` flows. This is the reverse of what we usually do with generics, but it is needed because `MatchingPropT` will inspect the upper bound of what it flows into, despite being a lower bound type itself.

Reviewed By: panagosg7

Differential Revision: D23224066

fbshipit-source-id: 40ee53de1eb8244342b158e2542bb8ecdf2475e1