[new-generics] Special-case support for `new this.constructor`
Summary:
This diff special-cases the `GetPropT` (when the prop is "constructor" and the bound is a class) and `ConstructorT` fields to preserve "genericness" across them. This is needed to support `new this.constructor()` having the type `this`: without it, the result of the call will have the non-generic type of the enclosing class, rather than the a generic type bounded by it like `this` does--so a program like this https://fburl.com/v2slf7wy would fail.

This admits all the existing unsoundness with `new this.constructor` (e.g. what if a subclass has a constructor with different arity), but is needed unless we want to ban `this.constructor` at all.

Reviewed By: panagosg7

Differential Revision: D23221893

fbshipit-source-id: 463454a9f2958ada99a5de10ae641964a0274ecc