[new-generics] Generic array spreads
Summary:
This diff builds support for spreading generic arrays, tuples, and iterables.

There are two main cases that we need to think about when spreading arrays and friends, which we handle in similar ways. The first is for `ResolveSpreadsToArray[Literal]`, where we take a sequence of arguments and spread arguments and resolve them to an array. The challenge with generics is to know when the resulting array is itself a generic. For example, this should be allowed:
```
function a<X: $ReadOnlyArray<number>>(x: X): X {
  return [...x]
}
```
because `[...x]` contains only elements that correspond to the allowed elements of `X`, whatever they are. However (h/t Panos) this should not be allowed, because tuples are subtypes of $ReadOnlyArray:
```
function b<X: $ReadOnlyArray<number>>(x: X): X {
  return [...x, ...x]
}
```
Also,
```
function c<X: $ReadOnlyArray<number>>(x: X): X {
  return [42, ...x]
}
```
should *not* be allowed, since `X` could be instantiated with `$ReadOnlyArray<21>`. (Unfortunately, Flow currently accepts this!) Likewise,
```
function d<X: [number, number]>(x: X): X {
  return [...x, ...x]
}
```
should not be allowed due to the length difference (similar to b).

To accomplish this, while we're computing the spread elements, we keep track (for each element, whether it's a spread element or a regular one) whether it originated in (i.e. from a generic tuple), or is itself, a spreaded generic. Then at the end, when we reconstruct an overall array type, we try to find any single generic ID that is satisfied by every element, and if so we re-wrap the output array type in a `GenericT` with that ID. With `ReadOnlyArray`s, we have the additional restriction that we only re-wrap if there has been exactly one spread, so as to not change the length.

A similar process occurs for `ResolveSpreadsToMultiflowCall` and similar. We again track the generic-ness of the argument sequence, and when any leftover arguments (which don't match a non-rest parameter) are flowed to the rest parameter of the function, if any, they're wrapped in a generic if an appropriate one exists.

Reviewed By: panagosg7

Differential Revision: D23267750

fbshipit-source-id: bd315d1f91ab51dc436bcb67b163f01e7a28f762