[new-generics] Use GenericT rather than RPolyTest to track escaped generics
Summary: This diff changes the generic-escape mechanism to look directly for GenericTs instead of just RPolyTest reasons to see if something has escaped. This has advantages in the case of spreads, because now when a spread of generics escapes (like `{...X, ...Y}`), we can point at both generics separately in different errors.

Reviewed By: panagosg7

Differential Revision: D23660616

fbshipit-source-id: eb8994bc82086f288a17b36fa8bb0a8d00d75681