Deduplicate file paths when denormalizing saved state paths
Summary:
This is basically the dual of D21233334 (https://github.com/facebook/flow/commit/d5c093805791c29c4f3ad6f858023cc010c1ca31).

When denormalizing file paths in saved state, we currently allocate a new string every time we encounter a path. Since saved state stores the same path in multiple different places, this is highly wasteful.

This diff introduces a cache used to memoize calls to the `denormalize_file_key` function. It is used everywhere except for the places where multiworkers are used to do some of the denormalization (we can't effectively use it for jobs conducted through multiworkers because the serialization process would lead to new allocations anyway).

I decided to leave the multiworkers as-is because I did performance testing both with this diff as presented as well as several trials with the multiworkers removed and the memoization applied to everything in denormalization. It didn't make much of a difference either to the reachable words or to the startup time. It appeared to make it very slightly slower, though, so I decided to leave it this way for now. Anyway, the data processed by the multiworkers is just dumped into the shared heaps and not retained by the main process in its OCaml heap, so duplication is unlikely to matter much to the overall memory usage there anyway.

This diff is both a memory and startup time win. The reduced allocations more than make up for the time spent doing hash table lookups.

Reviewed By: mroch

Differential Revision: D24285236

fbshipit-source-id: 0ad88c925437c49e780c96bbb5a36733222e2ac8