Use lists instead of maps for parsed and unparsed heaps in saved state
Summary:
Maps are expensive to construct and manipulate, and in this case we don't need the `O(log n)` lookups that they provide. Plus, we actually convert these maps to lists anyway during init.

This is a very small saved state size win, and a decent saved state init time win.

I may explore doing something similar with the dependency graph, but I don't think there are significant wins to be had by doing something similar with the other fields here, since the others are consumed as maps so we would have to convert them to maps anyway, or make some more invasive change to Flow. I think it's unlikely to be worth it. The one exception is the dependency graph, which I might try converting at some point.

Reviewed By: samwgoldman

Differential Revision: D24315881

fbshipit-source-id: ed2ff0e9487cd930660c6a5eacaca4d51caeea2b