Thread annot through expression
Summary:
To keep track of the annotation context, we introduce a `local_annot` param to `expression` (later in the stack we add it to even more functions in statement.ml).

This param will eventually be used as the "target" for inference, but we don't have the ability to select parts of an annotation just yet. To prepare for that eventuality, I made the type of `local_annot` a `unit option`. For the purposes of erroring, we only care that `local_annot` is `Some _`, but we do not care about what the actual annotation is.

When we need to start tracking the inference target all occurences of `()` passed to `local_annot` should be replaced by the actual target annotation.

Reviewed By: panagosg7

Differential Revision: D24283818

fbshipit-source-id: 7892a53d69f7f1f68e2026e185843702a4525476