Add a flag to require local inference annotations
Summary:
When this flag is set to true, Flow will emit an error on positions that would require annotations in a local inference system.

The analysis that finds these positions is hidden behind this flag, and implemented in the stack above this diff.

My strategy for implementing the analysis is to first complain on all parameters without annotations and then get more complete by threading information though the AST pass.

Reviewed By: panagosg7

Differential Revision: D24223438

fbshipit-source-id: 252433b3bd11e126617fc894920c1dde68f8b73f