remove JSON type checking tests
Summary:
The following diff removes support for checking code in JSON files.
Currently, this is possible under the "all" mode. However, support there
is broken in types-first, since we completely skip creating signatures
for JSON files.

This diff prepares the unit tests for that change by removing the ones that
rely on checking of JSON files.

Reviewed By: samwgoldman

Differential Revision: D24273232

fbshipit-source-id: 047b1b5263f94ea821e82948b1f2ebf5a3acbc99