parse json files directly into package heap
Summary:
This diff changes the behavior of checking JSON files and fixes support
for them in types-first.

Before this change, JSON files were parsed and stored in the AST heaps
along with regular .js files. The expectation was that JSON files would be
used in type checking when the '--all' flag was passed. Support for this
broke with types-first, where we failed to create a type signature entry for
JSON files.

With this change, the only JSON files we parse are `package.json`. The
ASTs for json files are no longer stored in shared memory, but immediately
used to create the package-heap from the parsing workers.

Changes in package.json files typically cause the server to restart, so we
don't treat these as regular rechecks. On a change in a package.json file
we reparse the file (like before) and update the package-heaps directly.
This happens in recheck_updates.ml and didn't need to be changed.

To support this change in saved-state, the structure of the state changed
since we no longer keep the package.json ASTs in the heaps. Instead,
we now store the package heaps directly. To save the state, since the
package heap API does not expose an iterator over its entries, we store
the set of package.json files in the server env, and iterate over those.

Reviewed By: nmote

Differential Revision: D24273233

fbshipit-source-id: eff3f9d98aaec3424357a7a16c9446df4d6bc1a3