Prepare hh_shared for new heap
Summary:
This diff slightly changes the C API for shared memory and the header format for
heap objects, in preparation for a more capable heap to support types-first 2.0.

First, the header format: Each heap object starts out with a header, which is a
single word that contains size information, so we know how to collect the heap
entry (how many bytes follow the header) and how to actually read the entry (how
big a buffer to allocate to decompress).

Before this change, we reserved only a single bit from the header, which is used
during garbage collection to distinguish headers from pointers. (Why? It's
complicated, and I don't want to get into it here, so just trust me.)

In the new heap, we want to reserve the full low order byte from headers. This
will give us 2 bits for GC and 6 bits for a "tag" which we will use to
distinguish between different kinds of objects -- currently all objects are of
the same kind: compressed and serialized OCaml values.

However, with only 56 bits to store the size information instead of 63, we can't
store the exact byte sizes anymore in the worst case. This diff also changes the
storage from byte sizes to word sizes. See inline comments in hh_store_ocaml for
more details.

For the C API, I have split add into store+add and get into get+deserialize.
Store writes into the heap and returns an address, which can be added to the
hash table in a separate step. Get returns an address registered with the
hash table, and deserialize reads the data from the heap.

This more granular API makes it possible to add heap entries *without*
registering them with the heap. This is currently unused, but is necessary to
create pointerful representations in the heap, where the hash table points to
entry A and entry A points to entry B (but the hash table does not point to B).

Note that this is only a small step toward supporting pointerful representations
in the heap. For example, the existing garbage collector is not powerful enough
to collect such representations, and will need to be modified.

Reviewed By: nmote

Differential Revision: D24214945

fbshipit-source-id: b6f5600fc935fa443480c6f8e89a21c04e66959c