Drop cache-line alignment of heap allocations
Summary:
Before this diff, each call to hh_alloc would pad out the requested size to a
full cache line. The motivation for this padding was to avoid cache line
bouncing when two threads write concurrently to the shared heap.

In practice, however, this is not a significant performance consideration. While
this scenario of two threads writing to the same cache line is possible, even
likely to occur, it's not part of a particularly hot section of code. In
practice, Flow will allocate a bunch of space during init and fill that space
from start to finish, only reading from then on.

The cache alignment is also significant for the garbage collector. The GC scans
the heap from start to finish. To perform this scan, we need to know how to get
from one heap entry to the next. Currently, we do this by reading the size of
the heap entry from its header, then padding that size out to the cache line.

Because of the GC interaction, this change also removes a small complication
from upcoming diffs, which add a new heap API. In the new API, we will write
more, smaller heap objects. It would be unfortunate to pad out each of those
objects, because the amount of space wasted to padding would become significant.

One possibility would be to allocate space for multiple objects, pad the
allocation itself, then write multiple objects into that space. In this scheme,
scanning the heap becomes difficult, since we no longer know how to get from one
heap object to the next.

We could work around that issue by writing a "pad" object at the end of each
allocation, which encodes the remaining size of the allocation in its header,
but by simply dropping the cache line alignment we can avoid the complication
entirely.

Reviewed By: nmote

Differential Revision: D24367161

fbshipit-source-id: 96f7e246d099e037251e9d1159257e69072cfe02