[new-generics] Handle KeysT, OptionalChainT, and ToStringT
Summary:
This diff handles some cases that were missed previously.

When a generic flows to an OptionalChainT, we now flow the bound of the generic to the OptionalChainT. This lets us avoid a catch-all case that would cause errors.

When a generic flows to ToStringT, we now flow the generic's bound to ToStringT and re-wrap the result in a generic. This is kind of iffy in general--we don't want to say that the result of applying an arbitrary transformation to a generic is itself a generic--but `ToStringT` seems to be primarily (exclusively?) used for object keys, which are always strings anyways. This rule allows programs like this to check:
```
function gn<TType>(jsEnum: {[TType]: string, ...}) {
  (Object.keys(jsEnum): Array<TType>);
}
```

When the *generic's* bound is a KeysT, we try to normalize it to the resulting type (i.e. union of string literals or `string`) before re-wrapping it as a generic and flowing that to an upper bound.

Reviewed By: panagosg7

Differential Revision: D24065857

fbshipit-source-id: 118ade23748414edf64d17c0b5e8193b8a8d14c8