[new-generics] Propagate generics through React
Summary:
This diff supports the React Kit for new generics. It's necessary for generics to flow through React, in order to make sure that generic lower bounds reach generic upper bounds successfully. Most React kit operations, however, are destructive--accessing information from within an incoming type. To handle this, this diff makes sure that the React kit can look through generics to see their upper bound, and perform whatever action is appropriate on that.

The exception is coerce_singleton which turns its input type into a singleton type (if possible) and returns it. In this case, the output (if successful) is re-wrapped as a generic. I haven't seen this make a difference, but this behavior is included to try to make the behavior of coerce_singleton more discrete.

Reviewed By: jbrown215

Differential Revision: D24227672

fbshipit-source-id: 38e4187f06a2f2e55e00d6143c9de1535a4aaf71