[new-generics] Special cases for type destructors
Summary: This diff handles a number of special cases for type destructors being applied to generics. To a large extent, this is just making sure that the implementation of type destructors is aware of generics--this implementation has a number of existing special cases, and now each of those special cases may also apply to a generic version. So for example, the special rules in `eval_destructor` for `OpenT` now has to consider `GenericT { OpenT }`. (We avoided a lot of these special cases in other places by using the `handle_generics` function to decide what to do, but type destructors can occur on the right as well as the left, and `handle_generics` is just concerned with generics on the left.)

Reviewed By: panagosg7

Differential Revision: D24227673

fbshipit-source-id: 183efc042a17e469550e73cb213be0dbaa420972