[new-generics] Shortcut empty generics using existing logic
Summary: This diff uses the existing `empty_success` function, which is used to decide when to short-circuit a flow of `EmptyT ~> u` based on what `u` is and what "flavor" the empty has, to similarly short-circuit flows of `GenericT { bound = EmptyT } ~> u`. This is important because previously we would see flows like `GenericT { EmptyT } ~> ElemT (...)`, which hit a lower-bound-wildcard case for ElemT, and then the operation being performed wouldn't know what to do with a `GenericT { EmptyT }` lower bound, because the `empty_success` logic prevented `EmptyT` from reaching this constraint in the first place.

Reviewed By: panagosg7

Differential Revision: D24227488

fbshipit-source-id: ef97e22e7a5eabf9e9e5f0366e9a6bab2ab54034