improve error message when attempting to bind classes
Summary:
When attempting to bind a class type, we emit an error of the format:

```
Cannot call `X.bind` because class `X` is not supported by unclassified use BindT. [incompatible-use]
```

We should never expose our internal representation of types or uses in our error messages. This improves the error message by adding a new `upper_kind` to the `EIncompatible` message.

Reviewed By: mvitousek

Differential Revision: D24335073

fbshipit-source-id: 13ad8d0d48b35a980bd5706e76217fb53a462d4e