Add variant to keep track of whether variable bindings have annotations on their declarators
Summary:
The `general` type of a binding is either the union of all of its lower bounds or an annotation if the binding had one.

Being able to distinguish between these two cases is important for local inference. An unannotated var should be allowed to widen, whereas an annotated one should not. Furthermore, an annotated var can have its annotation pushed down to future assignments whereas an unannotated one cannot.

In order to distinguish between an annotated and unannotated tvar, we introduce a variant that tracks this information.

This diff does not attempt to track every single place an annotation is added. I will introduce more annotation information as the stack grows using this mechanism. See D24373541 for an example.

Reviewed By: panagosg7

Differential Revision: D24373542

fbshipit-source-id: a237e97ceb0796eddccd9b61984535c59c08bfb0