Thread annot through object_
Summary:
When an object has an annotation:
```
const obj: X = {...}
```

we can push down the types of the properties in the annotations to the properties in the object itself. That means we can avoid requiring an annotation on the function param in cases like this:

```
const obj: {foo: number => number} = {foo: (x) => 3}
```

This diff threads the annot argument through argument to prepare to make this change, which happens in the next diff.

Reviewed By: panagosg7

Differential Revision: D24318236

fbshipit-source-id: 7d1b5d4c552122ee3e0e429dd4d5ea27eaacd355