Push annot from declarations to assignment expressions
Summary:
This diff is slightly more complicated than the rest in the stack.

When we see a variable declarator in `variable_decl`, we record whether the type we create for the variable is an annotation or not using the infrastructure from D24373542 (https://github.com/facebook/flow/commit/410ed27aeaac1c0dede017db7b62f0c1b5808e9c).

With that information, we can ensure that all variables being  assigned in an assignment have annotations and then push the annotations down to the rhs expression. We can't do that with how destructuring is implemented yet, so until then we just push down the fact that an annotation exists. Getting this fully working will take more work.

We are careful to only analyze the rhs assuming an annotation exists when _all_ the variables on the LHS are annotated. Note that the bindings in a pattern allow annotations and still parse, but they are unsupported. Until that parser error is actually raised, it is possible to introduce an annotation on a binding inside a destructuring assignment that is ignored in type checking but taken into account for pushing down annotations.

Reviewed By: panagosg7

Differential Revision: D24373541

fbshipit-source-id: bb9ae9534aac192051153dca6c44146332c9a7d5