Break Type_sig_collections -> Type_sig_heap dependency
Summary:
The collections are arrays, but their array-ness hides behind an interface. The
heap API has functions that take or return arrays, so this conversion happens
inside the collection.

It's awkward for the collection modules to depend on the heap. This diff instead
uses the `to_array` function to convert a collection to an array before writing
to the heap, then the `init` function to read the heap data back into the
collection.

Reviewed By: nmote

Differential Revision: D24336016

fbshipit-source-id: bf23c2ede252b91d6f6ffacf913162e3fce3f999